requirements.txt
absl-py==2.2.2
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
etils==1.12.2
filelock==3.18.0
fonttools==4.58.2
fsspec==2025.3.2
huggingface-hub==0.32.4
idna==3.10
importlib_resources==6.5.2
isodate==0.7.2
Jinja2==3.1.6
jsonpath-rw==1.4.0
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mlcroissant==1.0.17
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.5
opencv-python==4.11.0.86
packaging==25.0
pandas==2.2.3
pandas-stubs==2.2.3.250308
pillow==11.2.1
ply==3.11
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
rdflib==7.1.4
requests==2.32.3
safetensors==0.5.3
scipy==1.15.3
six==1.17.0
sympy==1.14.0
timm==1.0.15
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1
tqdm==4.67.1
types-pytz==2025.2.0.20250326
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
zipp==3.21.0
